querys & refactoring

+++ shop_items.sql
@@ -20,7 +20,7 @@ CREATE TABLE IF NOT EXISTS `shop_items` (
   `key` varchar(50) NOT NULL,
   `label` varchar(50) NOT NULL,
   `weight` int NOT NULL DEFAULT '1',
-  `price` decimal(20,6) NOT NULL,
+  `price` int NOT NULL,
   `offer` tinyint DEFAULT '0',
   PRIMARY KEY (`id`),
   UNIQUE KEY `key` (`key`)
